feat: move login to pages
