+++ client/src/pages/Login/index.js
@@ -1,0 +1,3 @@
+import component from "./Login";
+
+export default component;
